feature #23: migrate app to typescript
