+++ components/index.ts
@@ -1,3 +1,2 @@
 export { default as SearchBox } from './search-box'
 export { default as Card } from './card'
-export { default as TypeRelationships } from './type-relationships'
