+++ package.json
@@ -27,6 +27,7 @@
     "typescript": "^5.1.3"
   },
   "devDependencies": {
+    "@types/node": "20.3.1",
     "@types/react": "^18.2.12",
     "husky": "^8.0.3"
   },
