+++ pages/index.tsx
@@ -1,9 +1,9 @@
-import { SearchBox } from "../components";
+import { SearchBox } from '../components'
 
 export default function Home() {
   return (
     <div>
       <SearchBox />
     </div>
-  );
+  )
 }
