+++ theme/index.ts
@@ -1,0 +1,32 @@
+const theme = {
+  colors: {
+    alabaster: '#FAFAFA',
+    astral: '#327DA0',
+    aquamarine: '#A1DAD7',
+    azure: '#0080FF',
+    black: '#000000',
+    cobalt: '#003DB3',
+    gray: '#333333',
+    mercury: '#F7F7F7',
+    nobel: '#B3B3B3',
+    red: '#E60000',
+    rose: '#FFAA80',
+    sunset: '#FF5050',
+    transparent: 'inherit',
+    white: '#FFFFFF',
+  },
+  fonts: {
+    small0: '10px',
+    small4: '14px',
+    small6: '16px',
+    small8: '18px',
+    medium0: '20px',
+    medium2: '22px',
+    medium4: '24px',
+    medium6: '26px',
+    medium8: '28px',
+    big0: '30px',
+  },
+}
+
+export default theme
